requirements.txt
# Core dependencies
langchain
openai
anthropic
# For UNO API
unopython
